server/requirements.txt
asgiref==3.8.1
attrs==25.1.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cloudinary==1.42.2
colorama==0.4.6
cryptography==44.0.0
defusedxml==0.7.1
Django==5.1.5
django-ckeditor-5==0.2.15
django-cloudinary-storage==0.3.0
django-cors-headers==4.6.0
django-debug-toolbar==5.0.1
django-filter==24.3
django-js-asset==3.0.1
django-oauth-toolkit==3.0.1
djangorestframework==3.15.2
djangorestframework-csv==3.0.2
djangorestframework-xml==2.0.0
djangorestframework-yaml==2.0.0
drf-excel==2.5.2
drf-nested-routers==0.94.1
drf-redesign==0.4.0
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.2.1
et_xmlfile==2.0.0
idna==3.10
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jwcrypto==1.5.6
Markdown==3.7
oauthlib==3.2.2
openpyxl==3.1.5
packaging==24.2
pillow==11.1.0
pluggy==1.5.0
pycparser==2.22
pytest==8.3.4
pytest-django==4.9.0
python-dotenv==1.0.1
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.22.3
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
